- particles: no fading on classic TI-Nspire - idle: was broken on classic TI-Nspire (thanks calc84maniac) - idle: fix the interrupt mask (thanks calc84maniac) - idle: keep the timer interrupt masked if it was - sleep: use the timer in one shot mode to avoid missing a wrap
git-svn-id: https://www.unsads.com/scm/svn/nsptools@478

+++ Ndless/branches/3.0-tno/include/libndls.h
@@ -89,7 +89,7 @@ unsigned hwtype(void);
 #define IO_LCD_CONTROL IO(0xC000001C, 0xC0000018)
 
 #define SCREEN_BASE_ADDRESS     ADDR_(*(volatile unsigned*)0xC0000010)
-#define SCREEN_BYTES_SIZE       ({_scrsize();})
+#define SCREEN_BYTES_SIZE       ((int)({_scrsize();}))
 #define SCREEN_WIDTH            320
 #define SCREEN_HEIGHT           240
 #define BLACK                   0x0
